### 2024-04-18 #6643 @editorialbot
[PRE REVIEW]: piar: Price Index Aggregation in R
**Submitting author:** @marberts (<a href="http://orcid.org/0000-0003-2544-9480">Steve Martin</a>)
**Repository:** https://github.com/marberts/piar
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.8.0
