### 2024-04-18 #6643 @oliviaguest
@marberts can you make your licence file have some contents, e.g. copy-paste an available/relevant open licence, please? Also can you check the DOIs above too in case some missing? 😊
