### 2024-04-19 7e38b72
Added missing DOIs
