+++ paper/paper.bib
@@ -11,7 +11,8 @@
   title = {Price and Quantity Index Numbers},
   author = {Balk, B. M.},
   year = {2008},
-  publisher = {Cambridge University Press}
+  publisher = {Cambridge University Press},
+  doi = {10.1017/CBO9780511720758}
 }
 
 @book{vonderlippe2001,
@@ -25,14 +26,16 @@
   title = {Consumer price index manual: Theory and practice},
   author = {ILO, and IMF, and OECD, and Eurostat, and UN, and {World Bank}},
   year = {2020},
-  publisher = {International Monetary Fund}
+  publisher = {International Monetary Fund},
+  doi = {10.5089/9781484354841.069}
 }
 
 @book{ppimanual,
   title = {Producer price index manual: Theory and practice},
   author = {ILO, and IMF, and OECD, and Eurostat, and UN, and {World Bank}},
   year = {2004},
-  publisher = {International Monetary Fund}
+  publisher = {International Monetary Fund},
+  doi = {10.5089/9781589063044.069}
 }
 
 @article{saavedra2021,
