### 2024-04-19 #6643 @marberts
Thanks, @oliviaguest. I added the missing DOIs for the references.

I'm not sure what happened with the license, but this isn't the only case of the bot missing it (e.g., #6615). I'm pretty sure what I've done is correct for an R package but please let me know if I need to make a change.
