### 2024-05-01 3260f67
Added data frame method for elemental_index()
