### 2024-05-02 4c0209f
Fixed evaluation for weights
