### 2024-05-05 e2665ca
Added contrib to as_index()
