+++ man/as_index.Rd
@@ -11,11 +11,11 @@
 \usage{
 as_index(x, ...)
 
-\method{as_index}{default}(x, ..., chainable = TRUE)
+\method{as_index}{default}(x, ..., chainable = TRUE, contrib = FALSE)
 
-\method{as_index}{matrix}(x, ..., chainable = TRUE)
+\method{as_index}{matrix}(x, ..., chainable = TRUE, contrib = FALSE)
 
-\method{as_index}{data.frame}(x, ..., chainable = TRUE)
+\method{as_index}{data.frame}(x, ..., chainable = TRUE, contrib = FALSE)
 
 \method{as_index}{chainable_piar_index}(x, ..., chainable = TRUE)
 
@@ -29,6 +29,9 @@ as_index(x, ...)
 \item{chainable}{Are the index values in \code{x} period-over-period
 indexes, suitable for a chained calculation (the default)? This should be
 \code{FALSE} when \code{x} is a fixed-base (direct) index.}
+
+\item{contrib}{Should the index values in \code{x} be used to construct
+percent-change contributions? The default does not make contributions.}
 }
 \value{
 \code{as_index()} returns a price index that inherits from
