+++ man/contrib.Rd
@@ -3,11 +3,16 @@
 \name{contrib}
 \alias{contrib}
 \alias{contrib.piar_index}
+\alias{contrib2DF}
 \title{Extract percent-change contributions}
 \usage{
 contrib(x, ...)
 
 \method{contrib}{piar_index}(x, level = levels(x)[1L], period = time(x), ..., pad = 0)
+
+contrib2DF(x, level = levels(x)[1L], period = time(x), ...)
+
+contrib2DF(x, level = levels(x)[1L], period = time(x), ...)
 }
 \arguments{
 \item{x}{A price index, as made by, e.g., \code{\link[=elemental_index]{elemental_index()}}.}
@@ -16,7 +21,7 @@ contrib(x, ...)
 
 \item{level}{The level of an index for which percent-change contributions
 are desired, defaulting to the first level (usually the top-level for an
-aggregate index).}
+aggregate index). \code{contrib2DF()} can accept multiple levels.}
 
 \item{period}{The time periods for which percent-change contributions are
 desired, defaulting to all time periods.}
@@ -25,13 +30,17 @@ desired, defaulting to all time periods.}
 rectangular array when products differ over time. The default is 0.}
 }
 \value{
-A matrix of percent-change contributions with a column for each
-\code{period} and a row for each product (sorted) for which there are
+\code{contrib()} returns a matrix of percent-change contributions with a column
+for each \code{period} and a row for each product (sorted) for which there are
 contributions in \code{level}. Contributions are padded with \code{pad} to fit into a
 rectangular array when products differ over time.
+
+\code{contrib2DF()} returns a data frame with four columns: \code{period}, \code{level},
+\code{product}, and \code{value}.
 }
 \description{
-Extract a matrix of percent-change contributions from a price index.
+Extract a matrix or data frame of percent-change contributions from a price
+index.
 }
 \examples{
 prices <- data.frame(
@@ -54,6 +63,8 @@ index <- aggregate(index, pias, na.rm = TRUE)
 # Percent-change contributions for the top-level index
 
 contrib(index)
+
+contrib2DF(index)
 
 # Calculate EA contributions for the chained index
 
