+++ NAMESPACE
@@ -85,6 +85,7 @@ export(carry_backwards)
 export(carry_forward)
 export(chain)
 export(contrib)
+export(contrib2DF)
 export(elemental_index)
 export(expand_classification)
 export(interact_classifications)
