+++ NEWS.md
@@ -9,6 +9,8 @@ this function with pipes without needing to wrap the function in `with()`.
 - `as_index()` gains a new argument `contrib` to add contributions to
 pre-computed indexes.
 
+- Added `contrib2DF()` to extract percent-change contributions as a data frame.
+
 # Version 0.8.0
 
 ## Significant changes
