+++ NAMESPACE
@@ -36,6 +36,7 @@ S3method(chain,chainable_piar_index)
 S3method(chain,default)
 S3method(chain,direct_piar_index)
 S3method(contrib,piar_index)
+S3method(contrib2DF,piar_index)
 S3method(elemental_index,data.frame)
 S3method(elemental_index,default)
 S3method(elemental_index,numeric)
