### 2024-05-10 #6643 @marberts
I realize there's no editor assigned yet, but here's a few suggestions for reviewers.

schneiderpy (recently published a paper in JOSS about an R package for making concentration indexes in economics).
grahamjwhite (author of an R package for making price indexes).
arne-henningsen (author of an R package for making price indexes).
