### 2024-05-24 84b5dfa
Updated example in readme to use nse
