+++ README.md
@@ -71,12 +71,10 @@ head(ms_prices)
 #> 5 202001       B2       5  8.61
 #> 6 202001       B2       6  6.40
 
-elementals <- with(
-  ms_prices, 
-  elemental_index(
-    price_relative(price, period, product), 
-    period, business, na.rm = TRUE
-  )
+elementals <- elemental_index(
+  ms_prices,
+  price_relative(price, period, product) ~ period + business,
+  na.rm = TRUE
 )
 
 elementals
@@ -88,7 +86,7 @@ elementals
 #> B4    NaN       NaN       NaN 4.576286
 ```
 
-And an aggregation structure with the `aggregation_structure()`
+And an aggregation structure with the `as_aggregation_structure()`
 function.
 
 ``` r
@@ -103,13 +101,9 @@ head(ms_weights)
 #> 4       B4             12    622
 #> 5       B5             12    330
 
-pias <- with(
-  ms_weights,
-  aggregation_structure(
-    c(expand_classification(classification), list(business)),
-    weight
-  )
-)
+ms_weights[c("level1", "level2")] <- expand_classification(ms_weights$classification)
+
+pias <- as_aggregation_structure(ms_weights[c("level1", "level2", "business", "weight")])
 
 pias
 #> Aggregation structure for 5 elemental aggregates with 2 levels above the elemental aggregates 
