+++ README.Rmd
@@ -59,18 +59,16 @@ library(piar)
 
 head(ms_prices)
 
-elementals <- with(
-  ms_prices, 
-  elemental_index(
-    price_relative(price, period, product), 
-    period, business, na.rm = TRUE
-  )
+elementals <- elemental_index(
+  ms_prices,
+  price_relative(price, period, product) ~ period + business,
+  na.rm = TRUE
 )
 
 elementals
 ```
 
-And an aggregation structure with the `aggregation_structure()` function.
+And an aggregation structure with the `as_aggregation_structure()` function.
 
 ```{r}
 # Make an aggregation structure from businesses to higher-level
@@ -78,13 +76,9 @@ And an aggregation structure with the `aggregation_structure()` function.
 
 head(ms_weights)
 
-pias <- with(
-  ms_weights,
-  aggregation_structure(
-    c(expand_classification(classification), list(business)),
-    weight
-  )
-)
+ms_weights[c("level1", "level2")] <- expand_classification(ms_weights$classification)
+
+pias <- as_aggregation_structure(ms_weights[c("level1", "level2", "business", "weight")])
 
 pias
 ```
