### 2024-05-24 #6643 @arne-henningsen
> 👋 @arne-henningsen - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Perhaps I can do this. How much timer would I have to do this? And how should I submit my reviewer report? (Just post everything here?)
