### 2024-05-24 #6643 @crvernon
@arne-henningsen everything is done via GitHub in a thread like this.  So you can iteratively review with other reviewers and the author.  So you can post comments, issues, and pull requests as part of your review that link back to the main review thread.  Our turn around time is usually 4 weeks or so.  How does that sound?

If you are ok with the above, I'll notify you when I have two reviewers and we switch over to the full review thread.  At that time, you can automatically generate your checklist and have at it.

Just let me know and thanks for the quick response!
