### 2024-05-24 #6643 @schneiderpy
> 👋 @schneiderpy - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Perhaps I can do this too. At least I'll do my best ...
