### 2024-05-27 marberts/piar#5 @schneiderpy
License file
There are two License files (LICENSE and LICENSE.md) with the same content, which, in my opinion, makes one superfluous.

Part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
