### 2024-05-28 marberts/piar#5 @schneiderpy
Well, it seems that the final result on the dedicated website depends on the type of license .. For example, the `see` r package changed recently to MIT + file LICENSE while others (including my own pgk `concstats`) use a different type of license (e.g. the `dynamite` package). Maybe @crvernon can help to clarify this.
