### 2024-05-28 marberts/piar#5 @crvernon
Yes, it is common to include both in this case because of the way they are read in.  As long as both match (license type and language) you should be OK.

Thanks!
