### 2024-05-28 marberts/piar#5 @marberts
Cool. Both `LICENSE` and `LICENSE.md` agree with each other, so I'm going to close this issue. Thanks.
