### 2024-05-29 marberts/piar#6 @schneiderpy
Phrasing and typos
@marberts I have encountered some minor issues like phrasing, typos etc. that are not essential, but perhaps would "polish" the paper even more:
Possible typos: line 26 ... should read packages, line 50 ... should read businesses
Possible phrasing issue: In the paper example you say
> 47 The goal of this example is to illustrate some of the core features of piar by building a typical

however, in line 52 you say:

> 52 to measure inflation (although the values are purposely unrealistic).

This sounds a bit contradictory to me .. a rephrasing might help.

As well, it seems that you are - unintentionally - introducing kind of "artificial problems" using words like "hiccup" (line 80) or "wrinkle" (line 71). New users don't like to encounter problems or difficult stuff. Furthermore, you, as the author, already toke care of this. Perhaps rephrasing to "Next step is unpacking ... " or something similar would help.  All this from a marketing point of view ..

The subtitle of the reference ILO, 2020 should be "... Concepts and Methods" or do you refer to another reference?

(As part of [the JOSS review](https://github.com/openjournals/joss-reviews/issues/6781))
