### 2024-05-30 3fc76e7
Fixed typos/phrasing in paper
