+++ man/aggregate.piar_index.Rd
@@ -146,8 +146,8 @@ as.matrix(pias) \%*\% as.matrix(chain(index[letters[1:3]]))
 Balk, B. M. (2008). \emph{Price and Quantity Index Numbers}.
 Cambridge University Press.
 
-ILO, IMF, OECD, Eurostat, UN, and World Bank. (2020).
-\emph{Consumer Price Index Manual: Theory and Practice}.
+IMF, ILO, OECD, Eurostat, UNECE, and World Bank. (2020).
+\emph{Consumer Price Index Manual: Concepts and Methods}.
 International Monetary Fund.
 }
 \seealso{
