### 2024-05-30 marberts/piar#6 @marberts
Thanks for catching those typos and the incorrect citation. I've corrected both in the paper and corrected the reference to the CPI manual in the package documentation.

I've taken your suggestion about phrasing and removed the "artificial problems" in the example.

As for the contradictory part on line 52, I've made it more explicit that the data in the example are synthetic data, not real data. Although this doesn't affect the structure of the data or how the functions in the package get used, it means the resulting index series has unconventional values. This is by design to ensure that there is no risk of misinterpreting these data as being real data.

All these changes are in [3fc76e7](https://github.com/marberts/piar/commit/3fc76e73779a30fa37f963e2ec64ec8d3256f91f).
