### 2024-05-30 #6781 @marberts
Updated the paper in [3fc76e7](https://github.com/marberts/piar/commit/3fc76e73779a30fa37f963e2ec64ec8d3256f91f) to incorporate @schneiderpy's suggestions in https://github.com/marberts/piar/issues/6.
