### 2024-06-01 marberts/piar#7 @schneiderpy
Unexpected function results
@marberts Following the example in you JOSS paper I have encountered some unexpect results:

Example:
a) relatives <- with(ms_prices, price_relative(price, period, product)) gives the result as shown in the paper. It would be nice to add head(relatives) in the paper, so that the user can see what he should expect.
However, if I change the order of the attributes I get the following result:
 ```
relative <- with(ms_prices, price_relative(period, product, price))
head(relative)
1.14 <NA> 6.09 6.23 8.61  6.4
   1   NA   NA   NA   NA   NA
```
b) uninformative error msg
```
relatives <- with(ms_prices, price_relative(price, product))
Error in price_relative(price, product) :
  argument "product" is missing, with no default
```
c)
```
elementals <- with(ms_prices, elemental_index(relatives, business, period, na.rm = TRUE))
head(elementals)
Period-over-period price index for 4 levels over 4 time periods
              B1       B2       B3       B4
202001 1.0000000 1.000000 1.000000      NaN
202002 0.8949097      NaN 2.020004      NaN
202003 0.3342939      NaN 1.635335      NaN
202004       NaN 2.770456 0.537996 4.576286
```
still gives correct computation, however, in a different form
However:
```
elementals <- with(ms_prices, elemental_index(relatives, period, na.rm = TRUE))
head(elementals)
Period-over-period price index for 1 levels over 4 time periods
  202001   202002  202003  202004
1      1 1.716475 1.19045 1.49501
```
I suggest more defensive code if input of attributes and order of attributes are important for correct computations.

d)
head() and tail() functions for your object "pias" (see paper) doesn't result in a dataframe. Probably, because it is not of class dataframe.
```
class(pias)
[1] "piar_aggregation_structure"
```
This functionality would be nice if one has larger data sets.

As part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
