### 2024-06-02 marberts/piar#7 @marberts
Thanks, @schneiderpy.

## Argument matching

Are items a-c issues with the`price_relative()` and `elemental_index()` functions, or just how they're shown in the example? Most functions in R are sensitive to the presence and position of the arguments; e.g.,

```r
x = c(1:4, NA)
y = 5:1
all.equal(weighted.mean(x, y, na.rm = TRUE), weighted.mean(y, x, na.rm = TRUE))
#> [1] "'is.NA' value mismatch: 1 in current 0 in target"

y = mtcars
merge(y)
#> Error in merge.data.frame(y) : argument "y" is missing, with no default
```

Perhaps it would be clearer if I write the argument names in the function call. For example, in your example c it would then be clear that the business labels are taken as time periods and the time periods as elemental aggregates

```r
elementals <- with(
  ms_prices,
  elemental_index(
    relatives,
    period = business,
    ea = period,
    na.rm = TRUE
  )
)
```

or that there are no elemental aggregates supplied (so the default applies, grouping all price relatives into a single elemental aggregate as documented)

```r
elementals <- with(
  ms_prices,
  elemental_index(
    relatives,
    period = period,
    na.rm = TRUE
  )
)
```

As for more defensive code, the `price_relative()` and `elemental_index()` functions each coerce their inputs to the requires types (as documented) and check the length of the vector inputs early in the body of the function.

```r
price_relative(1:5, period = c(1, 1, 2, 2, 3), product = mtcars)
#> Error in xtfrm.data.frame(x) : cannot xtfrm data frames

elemental_index(1:5, period = c(1, 1, 2, 2))
#> Error in elemental_index.numeric(1:5, period = c(1, 1, 2, 2)) : input vectors must be the same length
```

(This is why your example a works. If `ms_prices$period` were formatted like `2020-01` instead of `202001` it wouldn't work.) I'm happy to consider any suggestions for a different approach or if there are cases where the argument checking is insufficient.

## Printing with head/tail

Your suggestion to include `head(relatives)` is good. I think it would be best to add it to the data frame to make it easier to see how it transforms prices.

```r
ms_prices$relative <- with(ms_prices, price_relative(price, period, product))

head(subset(ms_prices, business == "B1"))
#>    period business product price  relative
#> 1  202001       B1       1  1.14 1.0000000
#> 2  202001       B1       2    NA        NA
#> 3  202001       B1       3  6.09 1.0000000
#> 11 202002       B1       2  6.94        NA
#> 12 202002       B1       3  5.45 0.8949097
#> 20 202003       B1       2  2.32 0.3342939

# ... make elemental indexes
```

It's a bit trickier for `head(pias)`. Although you're right that this would be helpful for larger aggregation structures, both `head(x)`and `tail(x)` should return an object of the same type as `x`. Right now I have no subscripting method for aggregation structures as I'm not really sure what such a method should do (if anything). I usually do

```r
print(pias, max = 12)
#> Aggregation structure for 5 elemental aggregates with 2 levels above the elemental aggregates
#>  level1 level2 ea weight
#> 1      1     11 B1    553
#> 2      1     11 B2    646
#> 3      1     11 B3    312
 [ reached 'max' / getOption("max.print") -- omitted 2 rows ]
```

to peek into an aggregation structure (just as for a data frame or matrix ). One could also do `head(as.data.frame(pias))` to get back a data frame.
