### 2024-06-03 marberts/piar#7 @schneiderpy
@marberts The unexpected results just concern the price_relative() and elemental_index() functions using your example in the paper

```
relative <- with(ms_prices, price_relative(price, period, product))
head(relative)
 1  2  3  4  5  6
 1 NA  1  1  1  1

relative2 <- with(ms_prices, price_relative(period, product, price))
head(relative2)
1.14 <NA> 6.09 6.23 8.61  6.4
   1   NA   NA   NA   NA   NA
```
```
expect_true(expect_equal(relative, relative2))
Error: `relative` not equal to `relative2`.
Names: 'is.NA' value mismatch: 6 in current 0 in target
```

Both objects, "relative" and "relative2", are the computational results of price_relative(). Both objects might be subject of further computation in another function which leads to erronous results if the user does not respect the order/number of the attributes. My point is, that there is no error message popping up to alert the user of the missuse of the function.

```
relatives <- with(ms_prices, price_relative(price, product))
Error in price_relative(price, product) :
argument "product" is missing, with no default
```
.. argument "product" is not missing in this example (perhaps an error message saying " ... one argument is missing" would be more informative)

The same arguments (for defensive coding) hold for elemental_index()

> One could also do head(as.data.frame(pias)) to get back a data frame.

Is it possible to change the class(pias) to data.frame or as.data.frame? That would probably include a gain in flexibility since the user can apply futher packages e.g. ggplot, knitr or flextable.
