### 2024-06-04 marberts/piar#7 @marberts
Thanks for the clarification @schneiderpy.

I'm not sure how I could validate the inputs to `price_relative()` for your first point (but please suggest any ideas). I would need to know which columns in a table contain the correct variables for each of `price_relative()`'s arguments. The function makes sure that the first argument can be represented as a numeric and that the second two arguments can be represented as factors, but I'm not sure how I would check the content of these variables to determine if the inputs are valid or not.

As for `with(ms_prices, price_relative(price, product))`, that error message comes from R, not me. Indeed, we can expand the call to see that the `product` argument has not been supplied and that instead the values in the `product` column of `ms_prices` are used to denote time periods.

```r
print_call = function(f) {
  b = as.list(body(f))
  body(f) = as.call(c(b[1], quote(print(match.call())), b[-1]))
  f
}

price_relative = print_call(price_relative)

with(ms_prices, price_relative(price, product))
#> price_relative(x = price, period = product)
#> Error in price_relative(price, product): argument "product" is missing, with no default
```

It feels like you're expecting `with()` to match column names to argument names like `do.call()`. This brings me back to my suggestion to name the arguments to `price_relative()` and `elemental_index()` in the paper to make the calls to these functions explicit and not dependent on the order of the arguments. Another idea is to use a formula interface for non-standard evaluation instead of wrapping things in `with()`. Elemental indexes could then be made like

```r
elementals = elemental_index(
  ms_prices,
  price_relative(price, period, product) ~ period + business,
  na.rm = TRUE
)

# Or, with pipes.
elementals = ms_prices |>
  mutate(relative = price_relative(price, period = period, product = product)) |>
  elemental_index(relative ~ period + product, na.rm = TRUE)
```

It would be easy enough to add this functionality, and I'm happy to do so if you think it would make the presentation easier.

(A third option would be to dispense with NSE and write out `price_relative(ms_prices$price, period = ms_prices$period, product = ms_prices$product)`, but I find that a bit too verbose.)

As for your second point, changing the class of `pias` would be tricky. It's not really a tabular object. There are methods to coerce to/from a data frame, but the current data structure I think makes the most sense.

One option is to keep the aggregation data as a table rather than explicitly construct the aggregation structure object. Aggregating with a data frame that represents an aggregation structure works just as well (as documented).

```r
ms_weights[c("level1", "level2")] = expand_classification(ms_weights$classification)

pias = ms_weights[c("level1", "level2", "business", "weight")]
pias
#>   level1 level2 business weight
#> 1      1     11       B1    553
#> 2      1     11       B2    646
#> 3      1     11       B3    312
#> 4      1     12       B4    622
#> 5      1     12       B5    330

ms_index = aggregate(elementals, pias, na.rm = TRUE)
```

Again, let me know if you think this would make the example in the paper clearer.
