### 2024-06-05 marberts/piar#7 @schneiderpy
@marberts You can try and implement something simple to test for function arguments, e.g.

```
attributes(price_relative())
Error in price_relative() : argument "x" is missing, with no default
> attributes(price_relative(ms_prices$price))
Error in price_relative(ms_prices$price) :
  argument "period" is missing, with no default
> attributes(price_relative(ms_prices$price, ms_prices$period))
Error in price_relative(ms_prices$price, ms_prices$period) :
  argument "product" is missing, with no default
```
or, as you have already mentioned, something like this:

```
test <- function(a,b,c) {
  defined <- ls()
  passed <- names(as.list(match.call())[-1])

  if (any(!defined %in% passed)) {
    stop(paste("missing values for", paste(setdiff(defined, passed), collapse=", ")))
  }
  a+b+c
}

```
As part of the J[OSS review](https://github.com/openjournals/joss-reviews/issues/6781)
