### 2024-06-05 marberts/piar#7 @marberts
Thanks for the suggestion. I'm going to mark this issue as needing further thought. I've never contemplated changing the default behavior when function arguments are missing, so I need to look into it more.

I'm going to record a couple thoughts here so I don't forget.

1. Functions in `tidyverse` seem to rely on the default.

```r
dplyr::nth()  # Style guide uses nth() as an example for error messages.
#> Error in dplyr::nth() : argument "x" is missing, with no default

purrr::accumulate()
#> Error in purrr::accumulate() : argument ".f" is missing, with no default

tidyr::replace_na()
#> Error in replace_na.default() : argument "data" is missing, with no default

# Default for S3 dispatch with no default.
dplyr::mutate()
#> Error in UseMethod("mutate") : no applicable method for 'mutate' applied to an object of class "NULL"
```

2. Functions that change the default seem to do so when missing arguments are allowed (I don't use that approach in `piar`).

```r
fixest::feols()
#> Error in fixest::feols() : You must provide the argument 'data' (currently it is missing).
```
