### 2024-06-07 marberts/piar#8 @schneiderpy
Summery/Statement of need
@marberts you have mentioned in you Summery / Statement of need (both package documentation and in the paper) a "narrow" target audience of "statistical agencies".
Is it possible that a "broader" target audience might use or be interested in using the package too, such as, e.g., central banks, consumer protection agencies, and academic in general?
If so, it might be good to include this wider audience in the package documentation and paper.

As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
