### 2024-06-07 marberts/piar#9 @schneiderpy
Community guidelines
Lack of community **guidelines**

> Community guidelines[](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)
> There should be clear guidelines for third-parties wishing to:
>
> Contribute to the software
>
> Report issues or problems with the software
>
> Seek support

The package might benefit from a _Contributing_ file.

As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
