+++ README.Rmd
@@ -26,6 +26,13 @@ knitr::opts_chunk$set(
 
 Most price indexes are made with a two-step procedure, where period-over-period *elemental indexes* are first calculated for a collection of *elemental aggregates* at each point in time, and then aggregated according to a *price index aggregation structure*. These indexes can then be chained together to form a time series that gives the evolution of prices with respect to a fixed base period. This package contains a collection of functions that revolve around this work flow, making it easy to build standard price indexes, and implement the methods described by Balk (2008), von der Lippe (2001), and the CPI manual (2020) for bilateral price indexes.
 
+The tools in this package are designed to be useful for both researching new
+sources of data and methods to construct price indexes, and the regular
+production of price statistics. It is targeted towards economists,
+statisticians, and data scientists working at national statistical agencies,
+central banks, financial institutions, and in academia that want to measure and
+study the evolution of prices over time.
+
 ## Installation
 
 Get the stable version from CRAN.
