### 2024-06-09 feb6345
Added contributing.md (#9)
