+++ CONTRIBUTING.md
@@ -1,0 +1,45 @@
+# Contributing
+
+Contributions to `piar` are welcome and appreciated. Here are some ways that you
+can contribute.
+
+## Reporting bugs
+
+Open an issue with a small reproducible example of the bug. Feel free to
+include a PR with a patch.
+
+## Improving documentation
+
+### Typos and corrections
+
+You can fix typos, spelling mistakes, and make other small corrections to
+the documentation by editing
+the `roxygen2` comments in the corresponding `.R` (not `.Rd`) file. 
+
+### Adding documentation
+
+Please open an issue if you feel that any of the
+documentation is unclear or if certain functions don't have enough documentation.
+
+## Suggestions for improvement
+
+Open an issue if you have any suggestions for improvements (e.g., feature requests).
+
+## Seeking support
+
+Open an issue (or send me an email) if you're stuck and not able to find an
+answer in the documentation. (It may be a bug or problem with the documentation.)
+
+## Making a pull request
+
+When making a pull request, please:
+
+1. Fork the repository on github and make a new branch in your forked copy;
+
+2. Follow the [tidyverse style guide](https://style.tidyverse.org/);
+
+3. Use `roxygen2` for documentation;
+
+4. Use `testthat` for unit tests;
+
+5. Ensure your changes pass `R CMD check`.
