### 2024-06-09 marberts/piar#9 @marberts
Thanks for the suggestion. I've added `CONTRIBUTING.md` in [feb6345](https://github.com/marberts/piar/commit/feb63451a522d5ae1b6682fb8766ef9222553bec) with instructions for making contributions, reporting issues/bugs, and seeking support.
