### 2024-06-10 #6781 @marberts
@crvernon Here's a short update from me.

I've incorporated several suggestions from @schneiderpy:
- marberts/piar#6: corrected some typos and an incorrect reference, and made a couple parts of the paper clearer.
- marberts/piar#8: made the target audience for the package clearer in the paper and `README.md`.
- marberts/piar#9: added `CONTRIBUTING.md` to the repo.

A few ideas/suggestions came up in marberts/piar#6. Some of these need more thought (i.e., changing the default behavior of argument matching), but there's good suggestions in here to make the example in the paper a bit clearer.
