### 2024-06-11 marberts/piar#10 @schneiderpy
Repository - Clarification needed
Yesterday I was reviewing the suggestions mentioned [here](https://github.com/openjournals/joss-reviews/issues/6781#issuecomment-2158289330), in the main and default branch of the repository as mentioned in the initial review threat. However, I could not find any of them, including the status badge. I'm aware, that there's a different branch in the repo called JOSS. However, as far as I know, people use a different branch for the review with the paper.md file, and after the review most of them delete this branch.
JOSS review guide says:

>The paper may be in a short-lived branch which is never merged with the default, although if you do this, make sure this branch is created from the default so that it also includes the source code of your submission.

Since this procedure is, in general, based on a temporal branch, my concern is that the main branch will not reflect the JOSS review process.
To avoid further confusion and perhaps unjustified "acceptance-blockers" I would like to receive some clarification on this @crvernon

As part of the [review ](https://github.com/openjournals/joss-reviews/issues/6781)
