### 2024-06-12 marberts/piar#10 @marberts
@schneiderpy Here's how I'm planning to incorporate the changes from this review. The `joss` branch contains all the work done for the paper and all the changes that have come from your comments and suggestions so far. I will continue to make changes in this branch as the review progresses. I've kept it separate from the `main` branch for now as `main` holds the stable development version of the package, and feeds to R-Universe.

Once the review is done, I'll bump the version from the current 0.8.0 to either 0.8.1 (if there are no breaking changes) or 0.9.0 (if there are significant changes) and merge into the `main` branch. All the commits in the `joss` branch will then be reflected in the history for `main`. I'll then make a release for this new version (which will automatically make the Zenodo record) and push the new version to CRAN and conda forge.

Hope that answers your question.
