### 2024-06-12 a4774b5
Updated paper to include example of price relatives (#7)
