### 2024-06-12 marberts/piar#7 @marberts
I've added an example of `head(relatives)` in the paper in [a4774b5](https://github.com/marberts/piar/commit/a4774b59f2b595ff5273db9e4466ab35f36f9a75).
