### 2024-06-12 marberts/piar#7 @marberts
I decided to merge in some work I had adding NSE to `elemental_index()`. It doesn't change the example in the paper, but I feel it can make the function a bit easier to use.
