### 2024-06-12 marberts/piar#10 @crvernon
Thanks @marberts - @schneiderpy I believe this should satisfy your question?  Let me know if not.  Thanks!
