### 2024-06-12 marberts/piar#7 @schneiderpy
@marberts since argument order matters in price_relative() you should probably name your arguments, e.g.
`price_relative(price = price, period = period, product = product)
`
However, I don't know if this works, since you are calling `gpindex::back_period(period, product) `for the final step.

I do agree that `ms_prices$relative` is probably uncessary.

head(pias) still doesn't work as expected
