### 2024-06-12 #6781 @schneiderpy
> 👋 @marberts, @schneiderpy, and @arne-henningsen - it looks like you are making good progress. Could you each provide a short update to how things are going here in this thread? Thanks!

@crvernon I have almost done ..
From my point of view [Unexpected function results](https://github.com/marberts/piar/issues/7) seems to be the main issue so far
