### 2024-06-13 e463de2
Named arguments to price_relative() in paper (#7)
