### 2024-06-13 marberts/piar#7 @marberts
Thanks for the suggestions. I ended up re-writing the code snippets in the example to try and incorporate your comments and make them easier to read (the results are unchanged).

The `period` and `product` arguments are now named for `price_relative()`. The first argument to `price_relative()` is `x`, not `price`, so I leave it unnamed as two out of three named arguments means the call is unambiguous. (I use `x` instead of `price` so that I have the option to easily make this function generic in the future.) `gpindex::back_period()` won't cause any issues here.

Instead of explicitly making an aggregation structure object, `pias` is now a data frame. This means `head(pias)` works as expected, as does any other function expecting a data frame. `aggregate()` converts this data frame to an aggregation structure object in the background.

Lastly, I've removed the use of `with()` in the example. Instead I use the data frame method for `elemental_index()` and specify the inputs with a formula. This is similar to using `tapply()` or `lm()`. I think this is a bit easier to read, especially if there's no need to show `ms_prices$relative`.
