### 2024-06-13 #6781 @marberts
The code snippets in the example have been re-written to address the comments in marberts/piar#7.
