### 2024-06-13 marberts/piar#7 @schneiderpy
Trying to replicate your examples in the re-written paper (not in the documentation which are still the same, however, in my opinion, should work with named function arguments)

The first example gives the following error:
```
elementals <- ms_prices |>
 transform(
    relative = price_relative(price, period = period, product = product)
   ) |>
   elemental_index(relative ~ period + business, na.rm = TRUE)
Error in elemental_index(as.numeric(x), ...) :
  'list' object cannot be coerced to type 'double'
```
Now, I'll use part of the chunk to "test" for price_relative()
 ```
elementals <- ms_prices |>
   transform(
     relative = price_relative(price, period = period, product = product)
   )
 head(elementals)
  period business product price relative
1 202001       B1       1  1.14        1
2 202001       B1       2    NA       NA
3 202001       B1       3  6.09        1
4 202001       B2       4  6.23        1
5 202001       B2       5  8.61        1
6 202001       B2       6  6.40        1
```
.. and with different order of the function arguments
 ```
elementals2 <- ms_prices |>
   transform(
     relative = price_relative(price, product = product, period = period)
   )
 head(elementals2)
  period business product price relative
1 202001       B1       1  1.14        1
2 202001       B1       2    NA       NA
3 202001       B1       3  6.09        1
4 202001       B2       4  6.23        1
5 202001       B2       5  8.61        1
6 202001       B2       6  6.40        1
```
worked fine

Now, I'll use the example in the package documentation (and previous example in the paper which led us to this issue)
(with correct order)
 ```
elementals3 <- with(
   ms_prices,
   elemental_index(
     price_relative(price, period, product),
     period, business, na.rm = TRUE
   )
 )
 head(elementals3)
Period-over-period price index for 4 levels over 4 time periods
   202001    202002    202003   202004
B1      1 0.8949097 0.3342939      NaN
B2      1       NaN       NaN 2.770456
B3      1 2.0200036 1.6353355 0.537996
B4    NaN       NaN       NaN 4.576286
```
with changed order
 ```
elementals4 <- with(
   ms_prices,
   elemental_index(
     price_relative(price, product, period),
     period, business, na.rm = TRUE
   )
 )
 head(elementals4)
Period-over-period price index for 4 levels over 4 time periods
      202001    202002    202003    202004
B1 1.0000000 0.7853026       NaN       NaN
B2 1.0166877       NaN 0.3633362 1.3831875
B3 0.9466638 0.6067904 2.1869572 0.8454208
B4       NaN       NaN 0.2242169 1.7651714
```
It might be the case that you didn't change the function itself (in the script)

head(pias) now returns a data frame which might be useful if one might plot historical price index evolutions

Additional question: what do the first three rows indicate in `chain(ms_index)`?

part of the[ review](https://github.com/openjournals/joss-reviews/issues/6781)
