### 2024-06-14 marberts/piar#7 @marberts
## Documentation

> ...not in the documentation which are still the same, however, in my opinion, should work with named function arguments

You're right, that should change. (It seems I didn't resolve the conflicts when I merged that code into the `joss` branch :disappointed: .)

## Error

Which version of `piar` are you using to get this error?

```r
elementals <- ms_prices |>
 transform(
    relative = price_relative(price, period = period, product = product)
  ) |>
  elemental_index(relative ~ period + business, na.rm = TRUE)

#> Error in elemental_index(as.numeric(x), ...) :
#>   'list' object cannot be coerced to type 'double'
```

Version 0.8.0 doesn't have the functionality I added the other day. You'll need to use version 0.8.0.9004 which is in the `joss` branch. You can install it with

```r
# You'll need to have {pak} installed.
# install.packages("pak")
pak::pak("marberts/piar@joss")
```

## Old example

I've not made any changes to the `price_relative()` function. I simply updated the examples to name the arguments to this function and remove the use of `with()`to make the function call clearer. Indeed,

```r
elementals3 <- with(
  ms_prices,
  elemental_index(
    price_relative(price, period = period,  product = product),
    period, business, na.rm = TRUE
  )
)

elementals4 <- with(
  ms_prices,
  elemental_index(
    price_relative(price, product = product, period = period),
    period, business, na.rm = TRUE
  )
)

all.equal(elementals3, elementals4)
#> [1] TRUE
```

## Additional question

```
#> Fixed-base price index for 8 levels over 4 time periods
#>    202001    202002    202003    202004
#> 1       1 1.3007239 1.3827662 3.7815355
#> 11      1 1.3007239 1.3827662 2.1771866
#> 12      1 1.3007239 1.3827662 6.3279338
#> B1      1 0.8949097 0.2991629 0.4710366
#> B2      1 1.3007239 1.3827662 3.8308934
#> B3      1 2.0200036 3.3033836 1.7772072
#> B4      1 1.3007239 1.3827662 6.3279338
#> B5      1 1.3007239 1.3827662 6.3279338
```

The first three rows are the aggregated indexes and this is what `aggregate(elementals, pias)` is computing. In this example businesses B1, B2, and B3 belong to industry 11, and B4 and B5 belong to industry 12. All five of these businesses belong to industry 1. So the index for industry 1 gives the change in price in this industry, which is constructed as the arithmetic index of the change in price for all five businesses. In practice this is usually a Lowe or Young index, but that depends on how the weights in `pias` are made.
