+++ R/price_relative.R
@@ -24,7 +24,11 @@
 #' [`gpindex::outliers`] for methods to identify outliers with price relatives.
 #'
 #' @examples
-#' price_relative(1:6, rep(1:2, each = 3), rep(letters[1:3], 2))
+#' price_relative(
+#'   1:6,
+#'   period = rep(1:2, each = 3),
+#'   product = rep(letters[1:3], 2)
+#' )
 #'
 #' @export
 price_relative <- function(x, period, product) {
