### 2024-06-14 marberts/piar#7 @schneiderpy
Thank you for the explication of the additional question. It might be a good idea to document this in the package documentation and paper.

Well, the unexpected function results still persist. Now tested with the new version

```
 packageVersion("piar")
[1] ‘0.8.0.9005’
```

```
 elementals <- ms_prices |>
   transform(
     relative = price_relative(price, period = period, product = product)
   ) |>
   elemental_index(relative ~ period + business, na.rm = TRUE)

 head(elementals)
Period-over-period price index for 4 levels over 4 time periods
   202001    202002    202003   202004
B1      1 0.8949097 0.3342939      NaN
B2      1       NaN       NaN 2.770456
B3      1 2.0200036 1.6353355 0.537996
B4    NaN       NaN       NaN 4.576286
```
 ```
elementals2 <- ms_prices |>
   transform(
     relative = price_relative(price, product, period)
   ) |>
   elemental_index(relative ~ period + business, na.rm = TRUE)
 head(elementals2)
Period-over-period price index for 4 levels over 4 time periods
      202001    202002    202003    202004
B1 1.0000000 0.7853026       NaN       NaN
B2 1.0166877       NaN 0.3633362 1.3831875
B3 0.9466638 0.6067904 2.1869572 0.8454208
B4       NaN       NaN 0.2242169 1.7651714
 all.equal(elementals, elementals2)
[1] "Component “index”: Component 1: Mean relative difference: 0.03501192"
[2] "Component “index”: Component 2: Mean relative difference: 0.5224239"
[3] "Component “index”: Component 3: 'is.NA' value mismatch: 1 in current 2 in target"
[4] "Component “index”: Component 4: Mean relative difference: 0.5714595"
```
And the "standalone" function price_relative()

 ```
relative1 <- price_relative(ms_prices$price, ms_prices$period, ms_prices$product)
 relative2 <- price_relative(ms_prices$price, ms_prices$product, ms_prices$period)
 all.equal(relative1, relative2)
[1] "Names: 40 string mismatches"                        "'is.NA' value mismatch: 12 in current 12 in target"
```
I still think that you should name the arguments in the function(s), since we have already discovered that argument order is important and avoids erroneous results if no error message is present.

as part of the [review](https://github.com/openjournals/joss-reviews/issues/6781)
