### 2024-06-15 bd5bc5b
Clarity about aggregated indexes in example (#7)
