### 2024-06-15 marberts/piar#7 @marberts
> It might be a good idea to document this in the package documentation and paper.

Good suggestion; I've made it clear in the paper and readme (bd5bc5b86829d773f9765b89f63e7773a1a6a202).
