### 2024-06-15 marberts/piar#7 @marberts
I must admit that I'm confused about the unexpected result with `price_relative()` as I thought this was an issue about how the function is presented in the paper and documentation.

There's no ambiguity when calling `price_relative()` if the `period` and `product` arguments are named, as you demonstrated in your comment yesterday, and I thought your comment was about naming these arguments so that it's clear what the function expects as an input. (Version 0.8.0.9005 should have much more consistent use of named arguments through the documentation to make the function calls clearer.)

Changing the order of the `period` and `product` argument when they're unnamed shouldn't give the same result because of how R matches arguments in a function call: named arguments are matched first and then the remaining arguments are matched in the order they appear. I'm not really sure what's unexpected here.

> I still think that you should name the arguments in the function(s), since we have already discovered that argument order is important and avoids erroneous results if no error message is present.

Can you elaborate on this? Do you mean changing the signature of `price_relative()` to be

```r
price_relative = function(x, ..., period, product) {
  ...
}
```
so that `period` and `product` have to be named?
