### 2024-06-18 marberts/piar#7 @marberts
Thanks for taking the time to explain the issue again, @schneiderpy. I think I understand now.

The issue isn't that `price_relative(price, period, product)` should equal `price_relative(price, product, period)`, but rather that it is easy to inadvertently swap the `period` and `product` arguments in `price_relative()` (for which the function gives no warning) and this can ruin all subsequent calculations.

I'm not sure there is a way to correctly check the `period` and `product` arguments. Anything that can be coerced to a factor is acceptable for either argument, so I don't think it's possible to raise a warning if they've been reversed. (But please submit a PR if you have a method.) I think the way forward is to add the dot-dot-dots so that `period` and `product` need to be named. This is also a good opportunity to make `price_relative()` generic.

While we've been discussing `price_relative()`, there are a few more functions that would benefit from this approach.
