### 2024-06-19 marberts/piar#7 @schneiderpy
> The issue isn't that price_relative(price, period, product) should equal price_relative(price, product, period), but rather that it is easy to inadvertently swap the period and product arguments in price_relative() (for which the function gives no warning) and this can ruin all subsequent calculations.

That is exactly what is was trying to say, however, swaped function arguments should give an error not warning, since everything subsequent (result) is not worthwile ... However, it is up to you, if you would like to give the user the possibility to swap the function arguments and you solve this internally in you function (or in the function that will do the final computation) or if you will restrict the input to the correct order of the arguments and error if otherwise.

Of course you can try something like `price_relative(x, ...)`, however, I'm not sure that ellipses argument will change anything. They are usually used to give the possibility to add more arguments.

Question: Have you tryied to implement the named argument as mentioned in the previous commit?

> While we've been discussing price_relative(), there are a few more functions that would benefit from this approach.

That is true, but I don't wanted to scare you ;)

As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
