### 2024-06-19 9abb448
Try price_relative() as a generic
