+++ NAMESPACE
@@ -51,6 +51,8 @@ S3method(mean,piar_index)
 S3method(merge,chainable_piar_index)
 S3method(merge,direct_piar_index)
 S3method(merge,piar_index)
+S3method(price_relative,data.frame)
+S3method(price_relative,default)
 S3method(print,piar_aggregation_structure)
 S3method(print,piar_index)
 S3method(rebase,chainable_piar_index)
