### 2024-06-19 marberts/piar#7 @marberts
I've updated `price_relative()` in [9abb448](https://github.com/marberts/piar/commit/9abb448d4753e17f1d3deb607ee04fd18bad4cd4) so that `period` and `product` need to be named and cannot be passed to the function by position. (This is on the `nse` branch, not the `joss` branch yet,)

```r
elementals2 <- ms_prices |>
  transform(
    relative = price_relative(price, period = period, product = product)
  ) |>
  elemental_index(relative ~ period + business, na.rm = TRUE)

elementals3 <- ms_prices |>
  transform(
    relative = price_relative(price, product = product, period = period)
  ) |>
  elemental_index(relative ~ period + business, na.rm = TRUE)

all.equal(elementals2, elementals3)
#> [1] TRUE
```

Trying to use the arguments unnamed is an error.

```r
ms_prices |>
  transform(
    relative = price_relative(price, product, period)
 )

#> Error in price_relative.default(price, product, period) : argument "period" is missing, with no default
```

I don't believe there is anything else I can do on this issue. There is no way to further validate the `period` and `product` arguments to give an error if they're swapped. (Is `price_relative(1:4, c(1, 2, 1, 2), c(3, 3, 4, 4))` making price relatives for products 1 and 2 in periods 3 and 4, or products 3 and 4 in periods 1 and 2? Either interpretation could be right.) Adding the dot-dot-dots is the only way I know of to force arguments to be named.

I think the `carry_forward()`, `carry_backwards()`, `shadow_price()`, and `elemental_index()` are the only other functions that need to be updated in the same way as `price_relative()`.

Please let me know if these changes address your comment and are sufficient to close this issue, as it will be a bit of work to implement. They also introduce non-backwards compatible changes that will break at least some code that I know is in use.
