### 2024-06-20 marberts/piar#7 @schneiderpy
Perhaps you can try something like this:

```
price_relative <- function(x, ...)
UseMethod("price_relative")
 # you have the above part already
price_relative.default <- function(x, ...) {
  # extract arguments
  a <- list(...)
  # define correct order for parameters
  params <- c("price", "product") # or as.factor
  # check that all mandatory parameters are specified
  stopifnot(all(params %in% names(a)))
  # check order of mandatory parameters
  stopifnot(identical(names(a)[names(a) %in% params], params))
  # do calculations
 res <- x / x[gpindex::back_period(period, product)]
  names(res) <- as.character(product)
  res
}
```
Perhaps you want to change stopifnot with assertthat_ for more informative error messages

As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
