### 2024-06-20 marberts/piar#7 @marberts
I tried your suggestion but it doesn't pass `R CMD check` because the signature of the function doesn't have arguments `period` and `product` but the documentation does. These arguments clearly need to be documented because the function still expects them.

```
❯ checking Rd \usage sections ... WARNING
  Documented arguments not in \usage in Rd file 'price_relative.Rd':
    ‘period’ ‘product’

  Functions with \usage entries need to have the appropriate \alias
  entries, and all their arguments documented.
  The \usage entries must correspond to syntactically valid R code.
  See chapter ‘Writing R documentation files’ in the ‘Writing R
  Extensions’ manual.
```

Adding these arguments to `price_relative()` makes the extra checks redundant as omitting them will signal an error (e.g., argument "period" is missing, with no default) and there is no need to check the order because they must be passed by name, not position.

The check also raises a note about no visible bindings, but this can be fixed.

```
❯ checking R code for possible problems ... NOTE
  price_relative.default: no visible binding for global variable ‘period’
  price_relative.default: no visible binding for global variable
    ‘product’
    period product
```
