### 2024-06-20 marberts/piar#7 @schneiderpy
Well, you are right, documenting ellipses is difficult ..

Have you tried something like
`price_relative <- function(x,  price = price, product = product, ...)`?

The problem is to include the mandatory arguments price and product and to control for the order of both.
Since `price_relative()` is "just" evaluating function input, while `gpindex::back_period()` is computing the final result, what about spliting the task, say, `price_relative() `evaluates the input, while `gpindex::back_period()` (or `gpindex::offset_period()`) would control for the right order of the arguments? Would that be feasable?

As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
