### 2024-06-21 marberts/piar#7 @marberts
Adding the default arguments like you suggest makes a circular reference and gives a weird error if an argument is missing.

```r
# Add default arguments.
price_relative <- function(x, period = period, product = product, ...) {
  x <- as.numeric(x)
  period <- as.factor(period)
  product <- as.factor(product)

  if (different_length(x, period, product)) {
    stop("'x', 'period', and 'product' must be the same length")
  }

  res <- x / x[gpindex::back_period(period, product)]
  names(res) <- as.character(product)
  res
}

price_relative(ms_prices$price, ms_prices$period)
#> Error in is.factor(x) :
#>   promise already under evaluation: recursive default argument reference or earlier problems?
```

These could be renamed to something else, but then the function can behave in weird ways a use variables in the global environment as defaults. If both `period` and `product` are supplied then the function behaves the same as if there were no defaults. Your original point about inadvertently swapping these arguments in the function call still applies.

For your second question, there's no way to make `gpindex::back_period()` enforce how `price_relative()` matches arguments. The arguments passed to this function are evaluated in the local scope of the `price_relative()` function, after they are matched in the function call.
