### 2024-06-21 #6781 @marberts
We're almost at the 4-week mark on this review, and I want to make sure that I'm on track to address the items in the review so far.

@schneiderpy
1. Other than the part of marberts/piar#7 about argument matching, are there any outstanding items related to the unchecked items on [your checklist](https://github.com/openjournals/joss-reviews/issues/6781#issuecomment-2131437780)?
2. Is the part about argument matching in marberts/piar#7 an acceptance blocker? (I'm interpreting it more as a comment about how R evaluates function calls than something about the functionality in the package.)

@arne-henningsen
I don't see your checklist. Have you run into any issues with the package so far?

Thanks!
