### 2024-06-21 marberts/piar#7 @schneiderpy
Let's go back to the first idea, trying to implente

```
price_relative <- function(x, ...)
UseMethod("price_relative")
 # you have the above part already
price_relative.default <- function(x, ...) {
  # extract arguments
  a <- list(...)
  # define correct order for parameters
  params <- c("price", "product") # or as.factor
  # check that all mandatory parameters are specified
  stopifnot(all(params %in% names(a)))
  # check order of mandatory parameters
  stopifnot(identical(names(a)[names(a) %in% params], params))
  # do calculations
 res <- x / x[gpindex::back_period(period, product)]
  names(res) <- as.character(product)
  res
}
```
To pass CRAN check you could document the mandatory arguments within `@param ...` or in a new section called `@section Additional parameters:`

As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6781)
