### 2024-06-21 #6781 @schneiderpy
The 4-week mark is not an upper limit

> ✨ Please start on your review when you are able, and be sure to complete your review in the next six weeks, at the very latest

Responding to your questions:
1. I have updated the checklist. The remaining unchecked items are related (in my opinion) with the [#7](https://github.com/marberts/piar/issues/7)
2. I do agree, for example, in the below case of code chunk. This might be solved, in general, with more informative error/warning messages to guide the user.

```
relatives <- with(ms_prices, price_relative(price, product))
Error in price_relative(price, product) :
  argument "product" is missing, with no default
```
However, I do not agree in the case where the functionality of a function results in erronous results, without adverting the user, and, furthermore, are used in subsequent computations. In my opinion, the suggested solution [here](https://github.com/marberts/piar/issues/7#issuecomment-2182833413) seems promising. And, probably, adjusting the documentation will get rid of the warning message from the CRAN check.
