### 2024-06-21 marberts/piar#7 @marberts
Thanks for the suggestion, but I'm not a fan of this idea for a couple reasons.

1. Overloading the documentation for `...` to accommodate two arguments or adding a custom section for parameters is not idiomatic and likely to be confusing. It feels like a hack to get around `R CMD check`.
2. I'm not sure what purpose it serves to enforce the order of named arguments: `price_relative(x, period = a, product = b)` and `price_relative(x, product = b, period = a)` should both work. There is no scope to accidentally swap the `period` and `product` arguments when they're named. It's only a problem when passing arguments by position, i.e., the difference between `price_relative(x, a, b)` and `price_relative(x, b, a)` requires the user to remember that the second argument in `period` and the third argument is `product`.

I understand your comment here: the `period` and `product` arguments are sufficiently similar that it is easy to swap them in a call to `price_relative()` without raising an error. Unless you're looking at the documentation for the function, this may be an easy mistake to make. I think that's a fair comment, which is why I updated the examples in the paper/readme/help files to use named arguments to make it clear how the function works.

If a fix to the documentation is not sufficient, the only resolution I can see for this issues is to disable matching the `period` and `product` arguments by position in `price_relative()`. The idiomatic way to disable matching arguments by position in R is to force these arguments to be named by having `...` appear before them in the function's list of formal arguments. I already do this with a number of generic functions, so it wouldn't be out of place (and would be a good opportunity to make `price_relative() generic). I realize there are well-known reasons why the `...` can be problematic in R, and I'd prefer to use something like Python's keyword only arguments, but that's just how R works. I'm happy to make this change if it will address your concern and close the issue.

Right now I feel that we're just going back and forth and not really agreeing with the way R matches arguments in a function call. I don't think trying to change the semantics of the language is a way forward; this will just make the function harder to use.

In any case, let me know if making `period` and `product` "keyword only" for `price_relative()` addresses your comment. If so, I can make the change for this function and the others I listed above with similar arguments. If not, while I appreciate your comments and the discussion we've had on this topic, I don't think there's any further changes I can make based on this issue.
