### 2024-06-21 marberts/piar#7 @schneiderpy
My suggested solution is just an idea.
Any solution that resolves this isssue is welcome

> However, I do not agree in the case where the functionality of a function results in erronous results, without adverting the user, and, furthermore, are used in subsequent computations.
