+++ tests/testthat/test-stack.R
@@ -1,5 +1,7 @@
-epr1 <- elemental_index(1:8, rep(1:2, each = 4), rep(letters[1:4], 2),
-                        contrib = TRUE)
+epr1 <- elemental_index(1:8,
+  period = rep(1:2, each = 4), ea = rep(letters[1:4], 2),
+  contrib = TRUE
+)
 epr2 <- as_index(matrix(1:8, 4, 2, dimnames = list(letters[1:4], 3:4)))
 pias1 <- as_aggregation_structure(
   data.frame(level1 = "0", level2 = letters[1:4], weight = 1:4)
@@ -30,14 +32,20 @@ test_that("stacking and unstacking are opposite operations", {
 
 test_that("stacking returns the correct result", {
   epr3 <- stack(epr1, epr2)
-  expect_equal(as.matrix(epr3),
-               matrix(1:8, 4, 4, dimnames = list(letters[1:4], 1:4)))
+  expect_equal(
+    as.matrix(epr3),
+    matrix(1:8, 4, 4, dimnames = list(letters[1:4], 1:4))
+  )
   expect_identical(levels(epr3), levels(epr1))
   expect_identical(time(epr3), as.character(1:4))
-  expect_equal(contrib(epr3),
-               matrix(c(0, 4, 0, 0), 1, 4, dimnames = list("a.1", 1:4)))
-  expect_equal(as.matrix(stack(epr2, epr1)),
-               matrix(1:8, 4, 4, dimnames = list(letters[1:4], c(3:4, 1:2))))
+  expect_equal(
+    contrib(epr3),
+    matrix(c(0, 4, 0, 0), 1, 4, dimnames = list("a.1", 1:4))
+  )
+  expect_equal(
+    as.matrix(stack(epr2, epr1)),
+    matrix(1:8, 4, 4, dimnames = list(letters[1:4], c(3:4, 1:2)))
+  )
 })
 
 test_that("stack works with [", {
