### 2024-06-23 dfcab16
Consolidated formula code
