+++ man/impute_prices.Rd
@@ -76,7 +76,7 @@ elemental price indexes: 0 for a geometric index, 1 for an arithmetic index
 (the default), or -1 for a harmonic index. Other values are possible; see
 \code{\link[gpindex:generalized_mean]{gpindex::generalized_mean()}} for details.}
 
-\item{formula}{A two-part formula with prices on the left-hand
+\item{formula}{A two-sided formula with prices on the left-hand
 side. For \code{carry_forward()} and \code{carry_backward()}, the right-hand side
 should have time periods and products (in that order); for
 \code{shadow_price()}, the right-hand side should have time period, products, and
