### 2024-06-24 d4ba0a5
Stack now matches inputs
