+++ R/stack.piar_index.R
@@ -57,11 +57,14 @@ stack.direct_piar_index <- function(x, y, ...) {
 
 #' @export
 stack.piar_index <- function(x, y, ...) {
-  if (length(x$levels) != length(y$levels) || any(x$levels != y$levels)) {
+  if (length(x$levels) != length(y$levels) || !setequal(x$levels, y$levels)) {
     stop("'x' and 'y' must be indexes for the same levels")
   }
   if (any(x$time %in% y$time)) {
     stop("the same time periods cannot appear in both 'x' and 'y'")
+  }
+  if (any(x$levels != y$levels)) {
+    y <- y[x$levels]
   }
   x$index <- c(x$index, y$index)
   x$contrib <- c(x$contrib, y$contrib)
