### 2024-06-24 marberts/piar#7 @marberts
I've updated `price_relative()` in [d4ba0a5](https://github.com/marberts/piar/commit/d4ba0a5c4ae558796676da96ef3373cb40afb922) to require `period` and `product` to be named. This means that the order of these arguments doesn't matter, so they can't be easily mixed up, and the default error messages are clearer.

```r
all.equal(
  with(ms_prices, price_relative(price, period = period, product = product)),
  with(ms_prices, price_relative(price, product = product, period = period))
)
#> [1] TRUE

with(ms_prices, price_relative(price, period = period))
#> Error in price_relative.default(price, period = period) : argument "product" is missing, with no default

with(ms_prices, price_relative(price, product = product))
#> Error in price_relative.default(price, product = product) : argument "period" is missing, with no default
```

I've done the same thing for all the other functions with similar arguments, namely `elemental_index()`, `carry_forward()`, 'carry_backward()`, and `shadow_price()`. These functions are also all generic now, and having these argument be named will make it easier to extend these functions in the future.
