### 2024-07-03 d65122e
Added dots checking
