+++ tests/testthat/test-mean.R
@@ -48,7 +48,7 @@ test_that("aggregating over subperiods works", {
 test_that("aggregating with a window of 1 does nothing", {
   x <- as_index(matrix(1:9, 3))
   expect_equal(mean(x, window = 1), x)
-  expect_equal(mean(x, 9:1, window = 1, r = 0), x)
+  expect_equal(mean(x, weights = 9:1, window = 1, r = 0), x)
   expect_equal(mean(window(x, end(x))), window(x, end(x)))
 })
 
