+++ R/merge.piar_index.R
@@ -43,6 +43,7 @@ merge.direct_piar_index <- function(x, y, ...) {
 
 #' @export
 merge.piar_index <- function(x, y, ...) {
+  chkDots(...)
   if (length(x$time) != length(y$time) || any(x$time != y$time)) {
     stop("'x' and 'y' must be indexes for the same time periods")
   }

[…]
