### 2024-07-19 marberts/piar#11 @marberts
Add DOIs for references
1. Replace reference for von der Lippe's 2001 book with his more comprehensive 2007 book.
2. CRAN now publishes DOIs, so add these to the references.

Part of https://github.com/openjournals/joss-reviews/issues/6781.
