### 2024-07-19 65d3952
Added DOIs for referneces (#11)
