+++ paper/paper.bib
@@ -15,11 +15,12 @@
   doi = {10.1017/CBO9780511720758}
 }
 
-@book{vonderlippe2001,
-  title = {Chain Indices: A Study in Price Index Theory},
+@book{vonderlippe2007,
+  title = {Index Theory and Price Statistics},
   author = {{von der Lippe}, P.},
-  year = {2001},
-  publisher = {Federal Statistical Office, Wiesbaden}
+  year = {2007},
+  publisher = {Peter Lang},
+  doi = {10.3726/978-3-653-01120-3} 
 }
 
 @book{cpimanual,
@@ -56,6 +57,7 @@
   year = {2023},
   note = {R package version 0.6.0},
   url = {https://cran.r-project.org/package=IndexNumR},
+  doi = {10.32614/CRAN.package.IndexNumR}
 }
 
 @Manual{vonbergmann2021,
@@ -64,6 +66,7 @@
   title = {{cansim}: Accessing Statistics Canada Data Table and
     Vectors},
   url = {https://cran.r-project.org/package=cansim},
+  doi = {10.32614/CRAN.package.cansim}
 }
 
 @Manual{welsh2024,
@@ -79,6 +82,7 @@
   year = {2022},
   note = {R package version 0.1-8},
   url = {https://cran.r-project.org/package=micEconIndex},
+  doi = {10.32614/CRAN.package.micEconIndex}
 }
 
 @Manual{krausse2020,
@@ -87,6 +91,7 @@
   year = {2020},
   note = {R package version 0.3.2},
   url = {https://cran.r-project.org/package=hpiR},
+  doi = {10.32614/CRAN.package.hpiR}
 }
 
 @Manual{bialek2024,
@@ -95,6 +100,7 @@
   year = {2024},
   note = {R package version 0.1.9},
   url = {https://cran.r-project.org/package=PriceIndices},
+  doi = {10.32614/CRAN.package.PriceIndices}
 }
 
 @Manual{martin2023,
@@ -103,7 +109,7 @@
   year = {2023},
   doi = {10.5281/zenodo.10110159},
   url = {https://cran.r-project.org/package=rsmatrix},
-  note = {R package version 0.2.8},
+  note = {R package version 0.2.8}
 }
 
 @Manual{FHMCFSPI,
