+++ man/elemental_index.Rd
@@ -177,6 +177,8 @@ Cambridge University Press.
 IMF, ILO, OECD, Eurostat, UNECE, and World Bank. (2020).
 \emph{Consumer Price Index Manual: Concepts and Methods}.
 International Monetary Fund.
+
+von der Lippe, P. (2007). \emph{Index Theory and Price Statistics}. Peter Lang.
 }
 \seealso{
 \code{\link[=price_relative]{price_relative()}} for making price relatives for the same products over
