+++ README.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 
 <!-- badges: start -->
 [![CRAN status](https://www.r-pkg.org/badges/version/piar)](https://cran.r-project.org/package=piar)
-[![piar status badge](https://marberts.r-universe.dev/badges/piar)](https://marberts.r-universe.dev)
+[![piar status badge](https://marberts.r-universe.dev/badges/piar)](https://marberts.r-universe.dev/piar)
 [![Conda Version](https://img.shields.io/conda/vn/conda-forge/r-piar.svg)](https://anaconda.org/conda-forge/r-piar)
 [![R-CMD-check](https://github.com/marberts/piar/workflows/R-CMD-check/badge.svg)](https://github.com/marberts/piar/actions)
 [![codecov](https://codecov.io/gh/marberts/piar/graph/badge.svg?token=lHDHsGHsLd)](https://app.codecov.io/gh/marberts/piar)
@@ -24,7 +24,7 @@ knitr::opts_chunk$set(
 [![Mentioned in Awesome Official Statistics ](https://awesome.re/mentioned-badge.svg)](https://github.com/SNStatComp/awesome-official-statistics-software)
 <!-- badges: end -->
 
-Most price indexes are made with a two-step procedure, where period-over-period *elemental indexes* are first calculated for a collection of *elemental aggregates* at each point in time, and then aggregated according to a *price index aggregation structure*. These indexes can then be chained together to form a time series that gives the evolution of prices with respect to a fixed base period. This package contains a collection of functions that revolve around this work flow, making it easy to build standard price indexes, and implement the methods described by Balk (2008), von der Lippe (2001), and the CPI manual (2020) for bilateral price indexes.
+Most price indexes are made with a two-step procedure, where period-over-period *elemental indexes* are first calculated for a collection of *elemental aggregates* at each point in time, and then aggregated according to a *price index aggregation structure*. These indexes can then be chained together to form a time series that gives the evolution of prices with respect to a fixed base period. This package contains a collection of functions that revolve around this work flow, making it easy to build standard price indexes, and implement the methods described by Balk (2008), von der Lippe (2007), and the CPI manual (2020) for bilateral price indexes.
 
 The tools in this package are designed to be useful for both researching new
 sources of data and methods to construct price indexes, and the regular
@@ -116,4 +116,4 @@ Chiru, R., Huang, N., Lequain, M. Smith, P., and Wright, A. (2015). *The Canadia
 
 IMF, ILO, OECD, Eurostat, UNECE, and World Bank. (2020). *Consumer Price Index Manual: Concepts and Methods*. International Monetary Fund.
 
-von der Lippe, P. (2001). *Chain Indices: A Study in Price Index Theory*, Spectrum of Federal Statistics vol. 16. Federal Statistical Office, Wiesbaden.
+von der Lippe, P. (2007). *Index Theory and Price Statistics*. Peter Lang.
