+++ DESCRIPTION
@@ -14,9 +14,9 @@ Description: Most price indexes are made with a two-step procedure, where
     respect to a fixed base period. This package contains a collection of
     functions that revolve around this work flow, making it easy to build
     standard price indexes, and implement the methods described by
-    Balk (2008, ISBN:978-1-107-40496-0), von der Lippe (2001,
-    ISBN:3-8246-0638-0), and the CPI manual (2020, ISBN:978-1-51354-298-0)
-    for bilateral price indexes.
+    Balk (2008, <doi:10.1017/CBO9780511720758>), von der Lippe (2007,
+    <doi:10.3726/978-3-653-01120-3>), and the CPI manual (2020,
+    <doi:10.5089/9781484354841.069>) for bilateral price indexes.
 Depends: R (>= 4.0)
 Imports:
     stats,

[…]
