+++ README.Rmd
@@ -114,6 +114,6 @@ Balk, B. M. (2008). *Price and Quantity Index Numbers*. Cambridge University Pre
 
 Chiru, R., Huang, N., Lequain, M. Smith, P., and Wright, A. (2015). *The Canadian Consumer Price Index Reference Paper*, Statistics Canada catalogue 62-553-X. Statistics Canada.
 
-IMF, ILO, OECD, Eurostat, UNECE, and World Bank. (2020). *Consumer Price Index Manual: Concepts and Methods*. International Monetary Fund.
+IMF, ILO, Eurostat, UNECE, OECD, and World Bank. (2020). *Consumer Price Index Manual: Concepts and Methods*. International Monetary Fund.
 
 von der Lippe, P. (2007). *Index Theory and Price Statistics*. Peter Lang.
