### 2024-07-23 aba53a5
Better use of chkDots()
