### 2024-08-14 905df29
Made some clarifications to the paper (#12)
