### 2024-08-14 #6781 @crvernon
👋  @marberts - we are almost there!

I only need one change in the paper:

- [ ] LINE 141:  The "c" in "canada" should be capitalized.  This can be done in the bib file by using curly brackets around the characters you wish to maintain the formatting of.

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (make sure the list is correct and ordered the same as the paper). You may also add the authors' ORCID.

- [ ]  Please respond with the DOI of the archived version here along with the version number
