+++ paper/paper.bib
@@ -63,7 +63,7 @@
 @Manual{vonbergmann2021,
   year = {2021},
   author = {Jens {von Bergmann} and Dmitry Shkolnik},
-  title = {{cansim}: Accessing Statistics Canada Data Table and
+  title = {{cansim}: Accessing {Statistics Canada} Data Table and
     Vectors},
   url = {https://cran.r-project.org/package=cansim},
   doi = {10.32614/CRAN.package.cansim}
