+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
   piar: piar.html
-last_built: 2024-07-18T20:08Z
+last_built: 2024-08-14T15:07Z
 urls:
   reference: https://marberts.github.io/piar/reference
   article: https://marberts.github.io/piar/articles
