### 2024-08-14 #6781 @marberts
Thanks, @crvernon.

- Thanks for catching the typo in the reference list. I've corrected the .bib file.
- The reviewed version is released on GitHub. I bumped the version to 0.8.1 to reflect the changes from the review.
- The released version is archived on Zenodo: https://doi.org/10.5281/zenodo.1332329. The title and author list matches the paper, and the license is the same what's on GitHub.
