### 2024-08-22 #6781 @marberts
@crvernon I forgot to add that the released/archived version (0.8.1) is on the `main` branch, along with the corrected entry in the reference list, not the `joss` branch used for this review.
